+++ example/example_task_wait_in_different_ways.dart
@@ -5,25 +5,28 @@ import 'package:multitasking/multitasking.dart';
 Future<void> main() async {
   final tasks = [
     doSomeWorkWithError(100),
-    doSomeWork1(1, 200),
-    doSomeWorkWithError(300),
-    doSomeWork1(2, 400),
+    doSomeWork(1, 200),
+    doSomeWork(1, 300),
   ];
 
-  print('whenAny');
-  final firstTask = await whenAny(tasks);
-  print('$firstTask: ${firstTask.state.name}');
+  final progress = Progress((int percent) {
+    print('Waiting: $percent%');
+  });
+
+  print('whenAny()');
+  final firstTask = await whenAny(tasks, progress: progress);
+  print('${firstTask.toString()}: ${firstTask.state.name}');
   print('Tasks');
   print(tasks.map((e) {
     return '$e: ${e.state.name}';
   }).join(', '));
 
-  print('whenAll');
-  await whenAll(tasks);
+  print('whenAll()');
+  await whenAll(tasks, progress: progress);
 
   for (var i = 0; i < tasks.length; i++) {
     final task = tasks[i];
-    var s = '$task: ${task.state.name}';
+    var s = '${task.toString()}: ${task.state.name}';
     if (task.isCompleted) {
       s += ', result: ${task.result}';
     } else {
@@ -34,7 +37,7 @@ Future<void> main() async {
   }
 }
 
-Task<int> doSomeWork1(int n, int ms) {
+Task<int> doSomeWork(int n, int ms) {
   return Task.run(() async {
     await Future<void>.delayed(Duration(milliseconds: ms));
     return n;
@@ -48,7 +51,15 @@ Task<int> doSomeWorkWithError(int ms) {
   });
 }
 
-Future<void> whenAll<T>(List<Task<T>> tasks) async {
+Future<void> whenAll<T>(
+  List<Task<T>> tasks, {
+  Progress<int>? progress,
+}) async {
+  if (tasks.isEmpty) {
+    progress?.report(100);
+    return Future.value();
+  }
+
   final completer = Completer<void>();
   var count = 0;
   for (var i = 0; i < tasks.length; i++) {
@@ -59,7 +70,10 @@ Future<void> whenAll<T>(List<Task<T>> tasks) async {
       } catch (e) {
         //
       } finally {
-        if (++count == tasks.length) {
+        ++count;
+        final percent = count * 100 ~/ tasks.length;
+        progress?.report(percent);
+        if (count == tasks.length) {
           completer.complete();
         }
       }
@@ -69,7 +83,14 @@ Future<void> whenAll<T>(List<Task<T>> tasks) async {
   return completer.future;
 }
 
-Future<Task<T>> whenAny<T>(List<Task<T>> tasks) async {
+Future<Task<T>> whenAny<T>(
+  List<Task<T>> tasks, {
+  Progress<int>? progress,
+}) async {
+  if (tasks.isEmpty) {
+    throw ArgumentError('Task list must not be empty', 'tasks');
+  }
+
   final completer = Completer<Task<T>>();
   for (var i = 0; i < tasks.length; i++) {
     final task = tasks[i];
@@ -80,6 +101,8 @@ Future<Task<T>> whenAny<T>(List<Task<T>> tasks) async {
         //
       } finally {
         if (!completer.isCompleted) {
+          final percent = 1 * 100 ~/ tasks.length;
+          progress?.report(percent);
           completer.complete(task);
         }
       }
@@ -87,4 +110,20 @@ Future<Task<T>> whenAny<T>(List<Task<T>> tasks) async {
   }
 
   return completer.future;
+}
+
+class Progress<T> {
+  final FutureOr<void> Function(T) _callback;
+
+  final Zone _zone;
+
+  Progress(final FutureOr<void> Function(T) callback)
+      : _callback = callback,
+        _zone = Zone.current;
+
+  void report(T event) {
+    _zone.scheduleMicrotask(() {
+      _callback(event);
+    });
+  }
 }